feat(analytics): move IAnalyticsRepository, IAnalyticsProductSource, DTOs, and enums to Domain layer
