+++ backend/src/Anela.Heblo.Application/Features/Analytics/AnalyticsModule.cs
@@ -1,6 +1,7 @@
 using Anela.Heblo.Application.Features.Analytics.DashboardTiles;
 using Anela.Heblo.Application.Features.Analytics.Infrastructure;
 using Anela.Heblo.Application.Features.Analytics.Services;
+using Anela.Heblo.Domain.Features.Analytics;
 using Anela.Heblo.Application.Features.Analytics.UseCases.GetMarginReport;
 using Anela.Heblo.Application.Features.Analytics.UseCases.GetProductMarginAnalysis;
 using Anela.Heblo.Application.Features.Analytics.Validators;
