+++ backend/src/Anela.Heblo.Application/Features/Analytics/UseCases/GetBankStatementImportStatistics/GetBankStatementImportStatisticsRequest.cs
@@ -1,3 +1,4 @@
+using Anela.Heblo.Domain.Features.Analytics;
 using MediatR;
 
 namespace Anela.Heblo.Application.Features.Analytics.UseCases.GetBankStatementImportStatistics;
@@ -21,20 +22,4 @@ public class GetBankStatementImportStatisticsRequest : IRequest<GetBankStatement
     /// Type of date to use for grouping statistics
     /// </summary>
     public BankStatementDateType DateType { get; set; } = BankStatementDateType.ImportDate;
-}
-
-/// <summary>
-/// Enum defining which date field to use for statistics
-/// </summary>
-public enum BankStatementDateType
-{
-    /// <summary>
-    /// Use StatementDate (date of the bank statement)
-    /// </summary>
-    StatementDate,
-
-    /// <summary>
-    /// Use ImportDate (when statement was imported into our system)
-    /// </summary>
-    ImportDate
 }
