+++ backend/test/Anela.Heblo.Tests/Features/Analytics/DashboardTiles/InvoiceImportStatisticsTileTests.cs
@@ -1,6 +1,5 @@
 using Anela.Heblo.Application.Features.Analytics.DashboardTiles;
-using Anela.Heblo.Application.Features.Analytics.Infrastructure;
-using Anela.Heblo.Application.Features.Analytics.UseCases.GetInvoiceImportStatistics;
+using Anela.Heblo.Domain.Features.Analytics;
 using FluentAssertions;
 using Moq;
 using Xunit;
