+++ backend/test/Anela.Heblo.Tests/Features/Analytics/GetInvoiceImportStatisticsHandlerTests.cs
@@ -1,5 +1,5 @@
-using Anela.Heblo.Application.Features.Analytics.Infrastructure;
 using Anela.Heblo.Application.Features.Analytics.UseCases.GetInvoiceImportStatistics;
+using Anela.Heblo.Domain.Features.Analytics;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Primitives;
 using Moq;
